Uint: rename `HLIMBS` to `RHS_LIMBS` (#432)
I'm not even sure what "HLIMBS" was supposed to stand for... high limbs
perhaps?

It's a constant for the number of limbs in the `rhs` parameter, i.e. the
right-hand operand, so this renames it to `RHS_LIMBS`, which is much
easier to visually scan versus `LIMBS` than `HLIMBS` was.